CRM/backend/requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
mysqlclient==2.2.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
